Refactor: naming buses (#8)
* Refactor: QueryBusInterface to QueryBus

Renamed QueryBusInterface to QueryBus to follow the convention of interface naming in the project. This change improves readability and clarity of the code. The uses of the renamed interface in BaseController and InMemorySymfonyQueryBus classes have also been updated accordingly.

* Refactor: CommandBusInterface to CommandBus

Renamed CommandBusInterface to CommandBus and CommandBusSyncInterface to CommandBusSync to follow the naming convention for interfaces in the rest of the project. This improves the consistency and clarity of the code. Updated all corresponding uses of these interfaces in the InMemorySymfonyCommandBus, InMemorySymfonyCommandBusSync, and BaseController classes.

* Style: apply csfixer